pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "specdec-study"
version = "0.1.0"
description = "Experimentation environment for evaluating speculative decoding algorithms using vLLM"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "vllm>=0.6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "orjson>=3.9.0",
]

[project.scripts]
run-experiment = "specdec_study.run_experiment:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["specdec_study*"]
